refactor/4-move-anniversary-reminder-to-hr_addon_settings.py
